[REFACTOR] 하나의 DTO로 통합 #42
