+++ src/main/java/sopt/org/umbbaServer/global/util/fcm/FCMScheduler.java
@@ -9,9 +9,9 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
-import sopt.org.umbbaServer.domain.qna.dao.QnADao;
 import sopt.org.umbbaServer.global.exception.CustomException;
 import sopt.org.umbbaServer.global.exception.ErrorType;
+import sopt.org.umbbaServer.global.util.fcm.controller.dto.FCMPushRequestDto;
 
 @Slf4j
 @Service
@@ -24,8 +24,6 @@ public class FCMScheduler {
     @Value("${fcm.topic}")
     String topic;
 
-    private final QnADao qnADao;
-
 
     @Scheduled(cron = "0 0 23 * * ?")
     public void pushTodayQna() {
@@ -35,28 +33,24 @@ public class FCMScheduler {
                 () -> new CustomException(ErrorType.USER_HAVE_NO_QNALIST)
         );
         QnA lastQna = qnAList.get(qnAList.size()-1);*/
-            pushAlarm(PushRequest.sendTodayQna("section", "question"));
+            pushAlarm(FCMPushRequestDto.sendTodayQna("section", "question"));
         } catch (FirebaseMessagingException e) {
             log.error("푸시메시지 전송 실패!: {}", e.getMessage());
             throw new CustomException(ErrorType.FAIL_TO_SEND_PUSH_ALARM);
         }
     }
 
-    public void pushOpponentReply() {
+    public void pushOpponentReply(String question) {
         try {
-            log.info("오늘의 질문 알람 - 유저마다 보내는 시간 다름");
-        /*List<QnA> qnAList = qnADao.findQnASByUserId(userId).orElseThrow(
-                () -> new CustomException(ErrorType.USER_HAVE_NO_QNALIST)
-        );
-        QnA lastQna = qnAList.get(qnAList.size()-1);*/
-            pushAlarm(PushRequest.sendTodayQna("section", "question"));
+            log.info("상대방 답변 완료!");
+            pushAlarm(FCMPushRequestDto.sendOpponentReply(question));
         } catch (FirebaseMessagingException e) {
             log.error("푸시메시지 전송 실패!: {}", e.getMessage());
             throw new CustomException(ErrorType.FAIL_TO_SEND_PUSH_ALARM);
         }
     }
 
-    private void pushAlarm(PushRequest.PushMessage data) throws FirebaseMessagingException {
+    private void pushAlarm(FCMPushRequestDto.PushMessage data) throws FirebaseMessagingException {
 
         Notification notification = Notification.builder()
                 .setTitle(data.getTitle())
@@ -69,6 +63,7 @@ public class FCMScheduler {
                 .build();
 
         FirebaseMessaging.getInstance().send(message);
+        log.info("firebase messaging send() 성공 : {}", message);
     }
 
 }
