+++ src/main/java/sopt/org/umbbaServer/global/util/fcm/controller/FCMController.java
@@ -6,9 +6,9 @@ import org.springframework.web.bind.annotation.*;
 import sopt.org.umbbaServer.global.common.dto.ApiResponse;
 import sopt.org.umbbaServer.global.config.jwt.JwtProvider;
 import sopt.org.umbbaServer.global.exception.SuccessType;
+import sopt.org.umbbaServer.global.util.fcm.controller.dto.FCMPushRequestDto;
 import sopt.org.umbbaServer.global.util.fcm.FCMScheduler;
 import sopt.org.umbbaServer.global.util.fcm.FCMService;
-import sopt.org.umbbaServer.global.util.fcm.controller.dto.FCMNotificationRequestDto;
 
 import java.io.IOException;
 import java.security.Principal;
@@ -23,7 +23,7 @@ public class FCMController {
 
     @PostMapping
     @ResponseStatus(HttpStatus.OK)
-    public ApiResponse sendNotificationByToken(@RequestBody FCMNotificationRequestDto request, Principal principal) throws IOException {
+    public ApiResponse sendNotificationByToken(@RequestBody FCMPushRequestDto request, Principal principal) throws IOException {
 
         return ApiResponse.success(SuccessType.PUSH_ALARM_SUCCESS, fcmService.sendMessageTo(request, JwtProvider.getUserFromPrincial(principal)));
     }
