--- src/main/java/sopt/org/umbbaServer/global/util/fcm/controller/dto/FCMNotificationRequestDto.java
@@ -1,14 +1,0 @@
-package sopt.org.umbbaServer.global.util.fcm.controller.dto;
-
-import lombok.*;
-
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class FCMNotificationRequestDto {
-
-    private String targetToken;
-    private String title;
-    private String body;
-}
